app.py
#!/usr/bin/env python3
import aws_cdk as cdk
from netpulse.netpulse_stack import NetPulseStack

app = cdk.App()

NetPulseStack(
    app,
    "NetPulseStack",
    env=cdk.Environment(
        account=app.node.try_get_context("account"),
        region=app.node.try_get_context("region") or "us-east-1",
    ),
)

app.synth()
