requirements.txt
# CDK
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
boto3>=1.34.0
moto[dynamodb,stepfunctions,sns]>=5.0.0
